requirements.txt
Flask==2.2.5
gunicorn
requests
line-bot-sdk==2.4.1